fix(catalog): entity cards tab filtering
