+++ plugins/api-docs/api-report-alpha.md
@@ -1,0 +1,22 @@
+## API Report File for "@backstage/plugin-api-docs"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackstagePlugin } from '@backstage/frontend-plugin-api';
+import { ExternalRouteRef } from '@backstage/frontend-plugin-api';
+import { RouteRef } from '@backstage/frontend-plugin-api';
+
+// @public (undocumented)
+const _default: BackstagePlugin<
+  {
+    root: RouteRef<undefined>;
+  },
+  {
+    registerApi: ExternalRouteRef<undefined, true>;
+  }
+>;
+export default _default;
+
+// (No @packageDocumentation comment for this package)
+```
